+++ src/javascript/customization.js
@@ -10,14 +10,6 @@ License > https://github.com/Graviton-Code-Editor/Graviton-App/blob/master/LICEN
 */
 const tinycolor = require("tinycolor2");
 
-if (!fs.existsSync(highlights_folder)) {
-  //If the highlights folder doesn't exist
-  fs.mkdirSync(highlights_folder);
-  fs.copy(path.join(__dirname, "src", "Highlights"), highlights_folder, err => {
-    if (err) return console.error(err);
-  });
-}
-
 detectLanguages();
 const loadTheme = number => {
   i = number;